# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Netra NOC API",
    "description": "Network Operations Center API for monitoring and management",
    "version": "2.5.0",
    "contact": {
      "name": "Support",
      "email": "support@netra-noc.local"
    }
  },
  "servers": [
    {
      "url": "https://api.netra-noc.local",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "API Token for programmatic access"
      }
    },
    "schemas": {
      "Device": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "model": { "type": "string" },
          "code": { "type": "string", "pattern": "^[A-Z0-9]{2,4}$" },
          "type": {
            "type": "string",
            "enum": ["router", "firewall", "switch", "radio_wifi", "access_point", "server", "computer", "printer", "nvr", "cctv"]
          },
          "ip_address": { "type": "string" },
          "location": { "type": "string" },
          "status": { "type": "string", "enum": ["online", "warning", "offline", "unknown"] },
          "monitoring_method": { "type": "string", "enum": ["icmp", "snmp", "tcp", "api", "onvif"] },
          "tcp_port": { "type": "integer", "minimum": 1, "maximum": 65535, "nullable": true },
          "snmp_version": { "type": "string", "enum": ["v1", "v2c"] },
          "has_snmp_credentials": { "type": "boolean" },
          "cpuPercent": { "type": "number" },
          "memoryPercent": { "type": "number" },
          "trafficMbps": { "type": "number" },
          "latencyMs": { "type": "number" }
        }
      },
      "DeviceInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 2, "maxLength": 120 },
          "model": { "type": "string", "minLength": 2, "maxLength": 160 },
          "type": { "type": "string" },
          "code": { "type": "string", "pattern": "^[A-Z0-9]{2,4}$" },
          "ipAddress": { "type": "string" },
          "location": { "type": "string", "minLength": 2, "maxLength": 200 },
          "monitoringMethod": { "type": "string", "enum": ["icmp", "snmp", "tcp", "api", "onvif"] },
          "tcpPort": { "type": "integer", "minimum": 1, "maximum": 65535 },
          "snmpCommunity": { "type": "string", "minLength": 8, "writeOnly": true },
          "snmpVersion": { "type": "string", "enum": ["v1", "v2c"], "default": "v2c" }
        }
      },
      "ApplicationSettings": {
        "type": "object",
        "properties": {
          "organizationName": { "type": "string", "minLength": 2, "maxLength": 120 },
          "timezone": { "type": "string", "enum": ["Asia/Jakarta", "Asia/Makassar", "Asia/Jayapura", "UTC"] },
          "dashboardRefreshSeconds": { "type": "integer", "minimum": 10, "maximum": 300 },
          "defaultMonitoringMethod": { "type": "string", "enum": ["icmp", "tcp", "snmp"] }
        }
      },
      "Alert": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "deviceId": { "type": "string", "format": "uuid" },
          "severity": { "type": "string", "enum": ["critical", "warning", "info"] },
          "title": { "type": "string" },
          "status": { "type": "string", "enum": ["new", "investigating", "acknowledged", "resolved"] }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "username": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["admin", "operator", "viewer", "auditor"] }
        }
      },
      "NotificationChannel": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "type": { "type": "string", "enum": ["email", "webhook", "teams", "slack", "telegram"] },
          "name": { "type": "string" },
          "endpoint": { "type": "string" }
        }
      },
      "DeviceGroup": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "deviceCount": { "type": "integer" }
        }
      },
      "MaintenanceWindow": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "startTime": { "type": "string", "format": "date-time" },
          "endTime": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["scheduled", "active", "completed", "cancelled"] }
        }
      },
      "SLAPolicy": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "targetUptimePercent": { "type": "number" },
          "responseTimeMinutes": { "type": "integer" }
        }
      }
    }
  },
  "paths": {
    "/health/live": {
      "get": {
        "tags": ["Health"],
        "summary": "Server liveness probe",
        "responses": {
          "200": {
            "description": "Server is running",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "status": { "type": "string" } } }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": ["Health"],
        "summary": "Server readiness probe",
        "responses": {
          "200": {
            "description": "Server is ready",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "status": { "type": "string" }, "database": { "type": "string" } } }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login with username and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" },
                  "remember": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Logout user",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Logout successful"
          }
        }
      }
    },
    "/api/dashboard/overview": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get dashboard overview",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Dashboard overview data"
          }
        }
      }
    },
    "/api/devices": {
      "get": {
        "tags": ["Devices"],
        "summary": "List devices",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "name": "query", "in": "query", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "List of devices"
          }
        }
      },
      "post": {
        "tags": ["Devices"],
        "summary": "Create device",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "name": "X-CSRF-Token", "in": "header", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  { "$ref": "#/components/schemas/DeviceInput" },
                  { "required": ["name", "model", "type", "code", "ipAddress", "location", "monitoringMethod"] }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device created"
          }
        }
      }
    },
    "/api/devices/{id}": {
      "get": {
        "tags": ["Devices"],
        "summary": "Get a device",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Device details" }, "404": { "description": "Device not found" } }
      },
      "patch": {
        "tags": ["Devices"],
        "summary": "Update a device",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "X-CSRF-Token", "in": "header", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeviceInput" } } }
        },
        "responses": { "200": { "description": "Device updated" }, "404": { "description": "Device not found" } }
      },
      "delete": {
        "tags": ["Devices"],
        "summary": "Soft-delete a device",
        "security": [{ "sessionCookie": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "X-CSRF-Token", "in": "header", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "204": { "description": "Device disabled" }, "404": { "description": "Device not found" } }
      }
    },
    "/api/settings": {
      "get": {
        "tags": ["Settings"],
        "summary": "Get application settings",
        "security": [{ "sessionCookie": [] }],
        "responses": { "200": { "description": "Application settings" } }
      },
      "patch": {
        "tags": ["Settings"],
        "summary": "Update application settings",
        "security": [{ "sessionCookie": [] }],
        "parameters": [{ "name": "X-CSRF-Token", "in": "header", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApplicationSettings" } } }
        },
        "responses": { "200": { "description": "Application settings updated" }, "403": { "description": "Admin role required" } }
      }
    },
    "/api/notifications/channels": {
      "get": {
        "tags": ["Notifications"],
        "summary": "List notification channels",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "List of notification channels"
          }
        }
      },
      "post": {
        "tags": ["Notifications"],
        "summary": "Create notification channel",
        "security": [{ "sessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": { "type": "string", "enum": ["email", "webhook", "teams", "slack"] },
                  "name": { "type": "string" },
                  "endpoint": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Channel created"
          }
        }
      }
    },
    "/api/device-groups": {
      "get": {
        "tags": ["Device Groups"],
        "summary": "List device groups",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "List of device groups"
          }
        }
      },
      "post": {
        "tags": ["Device Groups"],
        "summary": "Create device group",
        "security": [{ "sessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created"
          }
        }
      }
    },
    "/api/topology": {
      "get": {
        "tags": ["Topology"],
        "summary": "Get network topology",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Network topology graph"
          }
        }
      }
    },
    "/api/sla/policies": {
      "get": {
        "tags": ["SLA"],
        "summary": "List SLA policies",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "List of SLA policies"
          }
        }
      },
      "post": {
        "tags": ["SLA"],
        "summary": "Create SLA policy",
        "security": [{ "sessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "targetUptimePercent": { "type": "number" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Policy created"
          }
        }
      }
    },
    "/api/reports": {
      "get": {
        "tags": ["Reports"],
        "summary": "List scheduled reports",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "List of reports"
          }
        }
      },
      "post": {
        "tags": ["Reports"],
        "summary": "Create scheduled report",
        "security": [{ "sessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "reportType": { "type": "string", "enum": ["sla", "uptime", "traffic"] },
                  "frequency": { "type": "string", "enum": ["daily", "weekly", "monthly"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report created"
          }
        }
      }
    },
    "/api/api-tokens": {
      "get": {
        "tags": ["API Tokens"],
        "summary": "List user's API tokens",
        "security": [{ "sessionCookie": [] }],
        "responses": {
          "200": {
            "description": "List of API tokens"
          }
        }
      },
      "post": {
        "tags": ["API Tokens"],
        "summary": "Create API token",
        "security": [{ "sessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "scopes": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Token created"
          }
        }
      }
    }
  }
}